IIEA-10491 Moved ParserConstants to shared package
